Move notebook and commands to plugins
